+++ CHANGELOG.md
@@ -18,6 +18,8 @@ AddActionMembers now adds ScriptMethods instead of ScriptProperties:
 Deprecated Recorder.Loop removed: use Recorder.Mode.Loop
 Recorder.FileType changed from method to write-only property
 
+Strip Gainlayers are now FloatArrayMember objects, see README for details
+
 ### Added
 
 - IRemote base class
@@ -43,6 +45,20 @@ Recorder.FileType changed from method to write-only property
 - Command.Save($filepath)
 - Command.StorePreset()
 - Command.RecallPreset()
+- AddAliasMembers meta function takes a hashtable `-MAP` of `alias = property`
+- Strip.Karaoke alias for Strip.K
+- Strip.EQGain1|EQGain2|EQGain3 with bass/low, mid/med, treble/high aliases, respectively
+- StripAudibility class with Strip.Audibility.Knob
+- Strip.Denoiser.Threshold
+- Strip.VAIO
+- Strip.Pitch, StripPitch class
+  - on
+  - drywet
+  - pitchvalue
+  - loformant
+  - medformant
+  - hiformant
+  - recallpreset($presetIndex)
 
 ### Changed
 
@@ -58,6 +74,10 @@ Recorder.FileType changed from method to write-only property
 - Recorder.Armstrip|Armbus -> BoolArrayMember: now have .Get()
 - Cast Recorder getters to types for consistency
 - Floats getters/setters now default to two decimal places.
+- Strip.Mono is now an alias for Strip.MC on virtual strips
+- Strip.AppMute|AppGain can now take an app index; see README for details
+- Strip Knob setters: explicit $arg types for consistency
+- Strip.Levels.Convert hidden and return type [float] -> [single] for naming consistency
 
 ### Fixed
 
@@ -69,6 +89,9 @@ Recorder.FileType changed from method to write-only property
 - vban route range (API documentation is incorrect)
 - vban.stream.sr: $this._port -> $this._sr
 - Recorder.channel values: 1..8 -> (2, 4, 6, 8)
+- Strip.Limit type [int] -> [float]
+- Missing closing parenthesis in AppMute value string
+- Strip Knob getters: `this.Getter_String('') -> [math]::Round($this.Getter(''), 2)`
 
 ## [3.3.0] - 2024-06-29